tests/wpt/mozilla/tests/incremental-layout/layout-phases-stacking-context.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
    setup({explicit_done:true});
    window.onload = () => {
        test((t) => {
            document.body.style.zIndex = 100;
            t.add_cleanup(() => {
                document.body.style.zIndex = null;
            });

            assert_array_equals(
              ServoTestUtils.forceLayout().phases,
              ["BuiltStackingContextTree", "BuiltDisplayList"]
            )
        }, "Updating body's z-index skips layout and overflow calculation");

        done();
    }
</script>
